Extract CxxOrchardShardTreeDelegate to the separate file
